INFRA-7403: Allow the logstash image to be updated
This allows the logstash image to be updated within a major version.
Elastic does not tag logstash with major versions so we use the Docker Hub API to get the latest relevant version.

In our case, Telemetry only use Logstash v7, so we search Docker Hub for images tagged with 7., filter and sort the results.